Crate variance

variance

Helpers for establishing the variance of lifetimes and type parameters.

Structs

Invariant
A marker for forcing T to be considered invariant.

InvariantLifetime
A marker for forcing 'id to be considered invariant.